+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [Feature] Flipper app loader errors
 - [Feature] Skip synchronization if file system on flipper not changed
+- [Feature] Pass flipper color from searching screen
 - [GitHub] Update github repo picture
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI